Add grid snapping to resize operation
